+++ src/assets/scss/index.scss
@@ -1,0 +1,7 @@
+//Main SASS file, if it gets too big perhaps break it out into multiple files to reduce strain on client
+//Loading order also matters as files will depend on others
+
+//First - The Recylables :D
+@import './variables';
+@import './mixins';
+
